🚚 move contribution mapping into contribution module
