+++ marketplace-infrastructure/src/database/repositories/contribution.rs
@@ -3,6 +3,7 @@ use diesel::prelude::*;
 use itertools::Itertools;
 use log::error;
 use marketplace_domain::*;
+use std::str::FromStr;
 
 impl ContributionProjectionRepository for Client {
 	fn find_by_id(
@@ -144,6 +145,32 @@ impl From<ContributionProjection> for models::Contribution {
 			duration: contribution.metadata.duration,
 			context: contribution.metadata.context,
 			type_: contribution.metadata.r#type,
+		}
+	}
+}
+
+impl From<models::Contribution> for ContributionProjection {
+	fn from(contribution: models::Contribution) -> Self {
+		Self {
+			id: contribution.id.parse().unwrap(),
+			contributor_account_address: contribution
+				.contributor_account_address
+				.map(|account| ContributorAccountAddress::from_str(account.as_str()).unwrap()),
+			project_id: contribution.project_id.parse().unwrap(),
+			issue_number: contribution.issue_number.parse().unwrap(),
+			status: contribution.status.parse().unwrap_or(ContributionStatus::Open),
+			// Safe to unwrap because the value stored can only come from an u8
+			gate: contribution.gate.try_into().unwrap(),
+			description: contribution.description,
+			external_link: contribution.external_link.map(|link| url::Url::parse(&link).unwrap()),
+			title: contribution.title,
+			metadata: ContributionProjectionMetadata {
+				difficulty: contribution.difficulty,
+				technology: contribution.technology,
+				duration: contribution.duration,
+				context: contribution.context,
+				r#type: contribution.type_,
+			},
 		}
 	}
 }
